Restructure fixtures, fix enum tests
